Bug 793925: part 7 - move sdk/test to test so that we can do require(test) without magic.
