+++ lib/test.js
